requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
werkzeug==3.0.0
zipp==3.17.0
python-dotenv==1.0.1
Flask-Cors==5.0.0
gunicorn==23.0.0
face-recognition==1.3.0